backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.2"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
alembic = "^1.13.3"
tenacity = "^9.0.0"
pyjwt = "^2.10.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.17"
bcrypt = "^4.2.0"
python-dotenv = "^1.1.0"
natasha = "^1.6.0"
yargy = "^0.16.0"
pymorphy2 = "^0.9.1"
python-docx = "^1.1.2"
pdfminer-six = "^20250416"
setuptools = "^78.1.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
